requirements.txt
flask
joblib
numpy
scikit-learn
xgboost
matplotlib
seaborn
flask-cors
mysql-connector-python